[BE] feat: Movie 댓글

+++ final-pjt-back/movies/views.py
@@ -55,6 +55,7 @@ def movie_detail(request, movie_pk):
 
 
 @api_view(['GET'])
+@permission_classes([IsAuthenticated])
 def comment_list(request):
     if request.method == 'GET':
         # comments = Comment.objects.all()
@@ -64,7 +65,8 @@ def comment_list(request):
 
 
 @api_view(['GET', 'DELETE', 'PUT'])
-def comment_detail(request, comment_pk):
+@permission_classes([IsAuthenticated])
+def comment_detail(request, comment_pk, username):
     # comment = Comment.objects.get(pk=comment_pk)
     comment = get_object_or_404(Comment, pk=comment_pk)
 
@@ -72,20 +74,23 @@ def comment_detail(request, comment_pk):
         serializer = CommentSerializer(comment)
         return Response(serializer.data)
 
-    elif request.method == 'DELETE':
-        comment.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        
+    if str(comment.user) == str(username):
+        if request.method == 'DELETE':
+            comment.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
-    elif request.method == 'PUT':
-        serializer = CommentSerializer(comment, data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data)
+        elif request.method == 'PUT':
+            serializer = CommentSerializer(comment, data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data)
 
     
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def comment_create(request, movie_pk):
     movie = get_object_or_404(Movie, pk=movie_pk)
     serializer = CommentSerializer(data=request.data)
